// src/generated/model/feeTemplateResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FeeTemplateResponseStatus } from './feeTemplateResponseStatus.ts';

export interface FeeTemplateResponse {
  id?: string;
  code?: string;
  name?: string;
  status?: FeeTemplateResponseStatus;
}